"Update Web3 onboarding and dependencies"
Updated the Web3 onboard initialisation to include a specific project ID when constructing the ledger module. In use-chain-setter.tsx, connectedChain property and logs have been added for debugging purposes.

The yarn.lock file was updated to reflect changes in dependencies, notably updates in walletconnect and other libraries for better performance and newer features. The @coinbase/wallet-sdk dependency was removed as it's no longer in use. Ledgerhq was updated to a newer version (1.1.0) to fix some security issues.

These changes improve the application's interface with Web3 and associated dependencies, thus providing more stable and secure operation.