Changed useRouter import path from 'next/router' to 'next/navigation'
In multiple files, the import path for useRouter has been updated from 'next/router' to 'next/navigation'. This change was made in response to an update in the Next.js library where the useRouter method has been moved to the 'next/navigation' directory. The affected files include various versions and strategies of Aave and Spark network or product pages, as well as the index file in the 'earn/dsr' directory and the app file in the root directory.