Commit message: Disable 'standalone' output in next.config.js
The 'standalone' output functionality in next.config.js has been commented out. This adjustment was necessary to fix a compatibility problem with the newer versions of Next.js as it seems the 'standalone' configuration is not working properly on the latest release. The change helps in avoiding errors during build processes.